Move deployment code for the Hibernate Search extension to a dedicated package
Just to avoid split packages, since we'll add some code into the root
package for this extension, but in the runtime module.
